#31 Removed FireAfterTransaction usage

+++ src/main/java/org/openwms/core/uaa/admin/impl/RoleServiceImpl.java
@@ -27,12 +27,13 @@ import org.openwms.core.uaa.api.RoleVO;
 import org.openwms.core.uaa.api.ValidationGroups;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.validation.annotation.Validated;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
-import java.util.Collections;
+import java.util.ArrayList;
 import java.util.List;
 
 import static java.lang.String.format;
@@ -51,12 +52,14 @@ class RoleServiceImpl implements RoleService {
     private final UserService userService;
     private final RoleMapper mapper;
     private final Translator translator;
+    private final ApplicationEventPublisher eventPublisher;
 
-    RoleServiceImpl(RoleRepository repository, UserService userService, RoleMapper mapper, Translator translator) {
+    RoleServiceImpl(RoleRepository repository, UserService userService, RoleMapper mapper, Translator translator, ApplicationEventPublisher eventPublisher) {
         this.repository = repository;
         this.userService = userService;
         this.mapper = mapper;
         this.translator = translator;
+        this.eventPublisher = eventPublisher;
     }
 
     /**
@@ -74,7 +77,7 @@ class RoleServiceImpl implements RoleService {
     @Override
     @Measured
     public @NotNull RoleVO findByPKey(@NotBlank String pKey) {
-        return mapper.convertToVO(getRole(pKey));
+        return mapper.convertToVO(findByPKeyInternal(pKey));
     }
 
     /**
@@ -84,7 +87,7 @@ class RoleServiceImpl implements RoleService {
     @Measured
     public @NotNull List<Role> findByNames(@NotNull List<String> roleNames) {
         var allRoles = repository.findByNameIn(roleNames);
-        return allRoles == null ? Collections.emptyList() : allRoles;
+        return allRoles == null ? new ArrayList<>(0) : allRoles;
     }
 
     /**
@@ -100,6 +103,7 @@ class RoleServiceImpl implements RoleService {
             throw new ResourceExistsException(format("Role with name [%s] already exists", role.getName()));
         }
         newRole = repository.save(newRole);
+        eventPublisher.publishEvent(new RoleEvent(newRole, RoleEvent.EventType.CREATED));
         LOGGER.debug("Created Role [{}]", newRole);
         return mapper.convertToVO(newRole);
     }
@@ -111,13 +115,15 @@ class RoleServiceImpl implements RoleService {
     @Measured
     @Validated(ValidationGroups.Modify.class)
     public @NotNull RoleVO save(@NotBlank String pKey, @NotNull(groups = ValidationGroups.Modify.class) @Valid RoleVO role) {
-        var existingRole = getRole(pKey);
+        var existingRole = findByPKeyInternal(pKey);
         existingRole.setName(role.getName());
         existingRole.setDescription(role.getDescription());
-        return mapper.convertToVO(repository.save(existingRole));
+        existingRole = repository.save(existingRole);
+        eventPublisher.publishEvent(new RoleEvent(existingRole, RoleEvent.EventType.MODIFIED));
+        return mapper.convertToVO(existingRole);
     }
 
-    private Role getRole(String pKey) {
+    private Role findByPKeyInternal(String pKey) {
         return repository.findBypKey(pKey).orElseThrow(() -> new NotFoundException(
                 translator.translate(ROLE_WITH_PKEY_NOT_EXIST, pKey),
                 ROLE_WITH_PKEY_NOT_EXIST,
@@ -130,7 +136,11 @@ class RoleServiceImpl implements RoleService {
     @Override
     @Measured
     public void delete(@NotBlank String pKey) {
-        repository.deleteByPKey(pKey);
+        var role = repository.findBypKey(pKey);
+        if (role.isPresent()) {
+            repository.deleteByPKey(pKey);
+            eventPublisher.publishEvent(new RoleEvent(role.get(), RoleEvent.EventType.DELETED));
+        }
     }
 
     /**
@@ -139,10 +149,11 @@ class RoleServiceImpl implements RoleService {
     @Override
     @Measured
     public @NotNull RoleVO assignUser(@NotBlank String pKey, @NotBlank String userPKey) {
-        var role = getRole(pKey);
+        var role = findByPKeyInternal(pKey);
         var user = userService.findByPKey(userPKey);
         role.addUser(user);
         role = repository.save(role);
+        eventPublisher.publishEvent(new RoleEvent(role, RoleEvent.EventType.MODIFIED));
         return mapper.convertToVO(role);
     }
 
@@ -152,10 +163,11 @@ class RoleServiceImpl implements RoleService {
     @Override
     @Measured
     public @NotNull RoleVO unassignUser(@NotBlank String pKey, @NotBlank String userPKey) {
-        var role = getRole(pKey);
+        var role = findByPKeyInternal(pKey);
         var user = userService.findByPKey(userPKey);
         role.removeUser(user);
         role = repository.save(role);
+        eventPublisher.publishEvent(new RoleEvent(role, RoleEvent.EventType.MODIFIED));
         return mapper.convertToVO(role);
     }
 }
